requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
backports.csv==1.0.7
beautifulsoup4==4.6.0
builtwith==1.3.3
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
decorator==4.0.11
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-cors-headers==3.0.2
django-import-export==1.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
django-simpleui==2.1.5
easygui==0.98.1
et-xmlfile==1.0.1
hyperlink==19.0.0
idna==2.6
imageio==2.1.2
incremental==17.5.0
jdcal==1.4.1
jieba==0.39
jsonpath==0.82
kiwisolver==1.1.0
lxml==4.3.3
matplotlib==3.1.1
moviepy==0.2.3.2
numpy==1.17.2
odfpy==1.4.0
opencv-python==4.1.1.26
openpyxl==2.6.2
parsel==1.5.1
Pillow==6.0.0
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
queuelib==1.5.0
requests==2.18.4
retrying==1.3.3
scipy==1.3.1
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
tqdm==4.11.2
urllib3==1.22
w3lib==1.20.0
wordcloud==1.5.0
xlrd==1.2.0
xlwt==1.3.0
you-get==0.4.1328
zope.interface==4.6.0
